WIP On Notifications

+++ Moasher.Application/Features/Users/Commands/DeleteUser/DeleteUserCommand.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Microsoft.EntityFrameworkCore;
 using Moasher.Application.Common.Constants;
 using Moasher.Application.Common.Exceptions;
 using Moasher.Application.Common.Interfaces;
@@ -13,10 +14,12 @@ public record DeleteUserCommand : IRequest<Unit>
 public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, Unit>
 {
     private readonly IIdentityService _identityService;
+    private readonly IMoasherDbContext _context;
 
-    public DeleteUserCommandHandler(IIdentityService identityService)
+    public DeleteUserCommandHandler(IIdentityService identityService, IMoasherDbContext context)
     {
         _identityService = identityService;
+        _context = context;
     }
     
     public async Task<Unit> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
@@ -38,6 +41,12 @@ public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, Unit>
         }
         
         await _identityService.DeleteUserAsync(user, cancellationToken);
+
+        var userNotifications = await _context.UserNotifications
+            .Where(n => n.UserId == request.Id)
+            .ToListAsync(cancellationToken);
+        _context.UserNotifications.RemoveRange(userNotifications);
+        await _context.SaveChangesAsync(cancellationToken);
         
         return Unit.Value;
     }
